fixup: relocate harden_tests to sibling mmap_tests.rs (Copilot)
Per Copilot's #1130 review, the inline `#[cfg(test)] mod harden_tests`
diverged from the project's sibling-tests convention used elsewhere
(`afxdp/flow_cache.rs` → `flow_cache_tests.rs`,
`afxdp/umem/mod.rs` → `tests.rs`). Moved to
`afxdp/umem/mmap_tests.rs` and replaced with the standard
`#[cfg(test)] #[path = "mmap_tests.rs"] mod harden_tests;` stub.

cargo test --release: 880 tests pass (878 baseline + 2 harden tests).
