+++ userspace-dp/src/afxdp/umem/mmap.rs
@@ -162,31 +162,7 @@ impl Drop for MmapArea {
     }
 }
 
+
 #[cfg(test)]
-mod harden_tests {
-    use super::*;
-
-    #[test]
-    fn new_rejects_zero_length() {
-        let err = match MmapArea::new(0) {
-            Ok(_) => panic!("expected error"),
-            Err(e) => e,
-        };
-        assert_eq!(err.kind(), io::ErrorKind::InvalidInput);
-    }
-
-    #[test]
-    fn new_rejects_overflowing_aligned_len() {
-        // usize::MAX cannot be rounded up to the next 2 MB boundary —
-        // checked_add catches it before mmap.
-        let err = match MmapArea::new(usize::MAX) {
-            Ok(_) => panic!("expected error"),
-            Err(e) => e,
-        };
-        let msg = format!("{err}");
-        assert!(
-            msg.contains("hugepage alignment"),
-            "expected hugepage-alignment error, got: {msg}",
-        );
-    }
-}
+#[path = "mmap_tests.rs"]
+mod harden_tests;
